feat: add tests for UserRepository & fix bugs

+++ node/src/main/resources/db/changelog/v-1.0/data/002-create-gifts-table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS gifts (
-    id                INTEGER            NOT NULL GENERATED ALWAYS AS IDENTITY,
+    id                BIGINT            NOT NULL GENERATED ALWAYS AS IDENTITY,
     name_gift         VARCHAR(64)        NOT NULL                             ,
     creator_id        BIGINT             NOT NULL                             ,
     status            VARCHAR(16)        NOT NULL                             ,
